+++ solutions/p26/p26.mojo
@@ -1,9 +1,9 @@
-from gpu import thread_idx, block_idx, block_dim, lane_id
-from gpu.host import DeviceContext
-from gpu.primitives.warp import shuffle_xor, prefix_sum, WARP_SIZE
+from std.gpu import thread_idx, block_idx, block_dim, lane_id
+from std.gpu.host import DeviceContext
+from std.gpu.primitives.warp import shuffle_xor, prefix_sum, WARP_SIZE
 from layout import Layout, LayoutTensor
-from sys import argv
-from testing import assert_equal, assert_almost_equal
+from std.sys import argv
+from std.testing import assert_equal, assert_almost_equal
 
 
 comptime SIZE = WARP_SIZE
@@ -65,7 +65,7 @@ def butterfly_parallel_max[
         # Start with half the warp size and reduce by half each step
         offset = WARP_SIZE // 2
         while offset > 0:
-            max_val = max(max_val, shuffle_xor(max_val, offset))
+            max_val = max(max_val, shuffle_xor(max_val, UInt32(offset)))
             offset //= 2
 
         # All threads now have the maximum value across the entire warp
@@ -103,10 +103,10 @@ def butterfly_conditional_max[
         # Butterfly reduction for both maximum and minimum: dynamic for any WARP_SIZE
         offset = WARP_SIZE // 2
         while offset > 0:
-            neighbor_val = shuffle_xor(current_val, offset)
+            neighbor_val = shuffle_xor(current_val, UInt32(offset))
             current_val = max(current_val, neighbor_val)
 
-            min_neighbor_val = shuffle_xor(min_val, offset)
+            min_neighbor_val = shuffle_xor(min_val, UInt32(offset))
             min_val = min(min_val, min_neighbor_val)
 
             offset //= 2
@@ -207,7 +207,7 @@ def warp_partition[
         # Butterfly reduction to get total across the warp: dynamic for any WARP_SIZE
         offset = WARP_SIZE // 2
         while offset > 0:
-            warp_left_total += shuffle_xor(warp_left_total, offset)
+            warp_left_total += shuffle_xor(warp_left_total, UInt32(offset))
             offset //= 2
 
         # Phase 4: Write to output positions
@@ -231,7 +231,7 @@ def test_butterfly_pair_swap() raises:
 
         with input_buf.map_to_host() as input_host:
             for i in range(SIZE):
-                input_host[i] = i
+                input_host[i] = Float32(i)
 
         input_tensor = LayoutTensor[dtype, layout, ImmutAnyOrigin](input_buf)
         output_tensor = LayoutTensor[dtype, layout, MutAnyOrigin](output_buf)
@@ -253,10 +253,10 @@ def test_butterfly_pair_swap() raises:
         for i in range(SIZE):
             if i % 2 == 0:
                 # Even positions get odd values
-                expected_buf[i] = i + 1
+                expected_buf[i] = Float32(i + 1)
             else:
                 # Odd positions get even values
-                expected_buf[i] = i - 1
+                expected_buf[i] = Float32(i - 1)
 
         with output_buf.map_to_host() as output_host:
             print("output:", output_host)
@@ -276,9 +276,9 @@ def test_butterfly_parallel_max() raises:
 
         with input_buf.map_to_host() as input_host:
             for i in range(SIZE):
-                input_host[i] = i * 2
+                input_host[i] = Float32(i * 2)
             # Make sure we have a clear maximum
-            input_host[SIZE - 1] = 1000.0
+            input_host[SIZE - 1] = Float32(1000.0)
 
         input_tensor = LayoutTensor[dtype, layout, ImmutAnyOrigin](input_buf)
         output_tensor = LayoutTensor[dtype, layout, MutAnyOrigin](output_buf)
@@ -318,9 +318,9 @@ def test_butterfly_conditional_max() raises:
             for i in range(SIZE_2):
                 if i < 9:
                     values = [3, 1, 7, 2, 9, 4, 8, 5, 6]
-                    input_host[i] = values[i]
+                    input_host[i] = Float32(values[i])
                 else:
-                    input_host[i] = i % 10
+                    input_host[i] = Float32(i % 10)
 
         input_tensor = LayoutTensor[dtype, layout_2, ImmutAnyOrigin](input_buf)
         output_tensor = LayoutTensor[dtype, layout_2, MutAnyOrigin](output_buf)
@@ -376,7 +376,7 @@ def test_warp_inclusive_prefix_sum() raises:
 
         with input_buf.map_to_host() as input_host:
             for i in range(SIZE):
-                input_host[i] = i + 1
+                input_host[i] = Float32(i + 1)
 
         input_tensor = LayoutTensor[dtype, layout, ImmutAnyOrigin](input_buf)
         output_tensor = LayoutTensor[dtype, layout, MutAnyOrigin](output_buf)
@@ -422,7 +422,7 @@ def test_warp_partition() raises:
             # Create: [3, 7, 1, 8, 2, 9, 4, 6, ...]
             test_values = [3, 7, 1, 8, 2, 9, 4, 6, 0, 10, 3, 11, 1, 12, 4, 13]
             for i in range(SIZE):
-                input_host[i] = test_values[i % len(test_values)]
+                input_host[i] = Float32(test_values[i % len(test_values)])
 
         input_tensor = LayoutTensor[dtype, layout, ImmutAnyOrigin](input_buf)
         output_tensor = LayoutTensor[dtype, layout, MutAnyOrigin](output_buf)
